Add a CalleeSavedRegisterUsage compiler benchmark
The frame pointer is the single available for HotSpot callee-saved register
and this is available only when the compilation is run under the
-XX:-PreserveFramePointer option. This register allows the compiler do not
spill a register used for function calling (parameters passing or the result
returning) and to reload the register from the frame pointer instead what may
lead to a significant performance gain. But what does this "significant" mean?
The goal of the benchmark is to measure and get the numbers.